Add artsy shipping fulfillment option types
