+++ src/schema/v2/order/exchangeErrorHandling.ts
@@ -1,4 +1,4 @@
-import { ORDER_MUTATION_FLAGS } from "./sharedTypes/sharedOrderTypes"
+import { ORDER_MUTATION_FLAGS } from "./types/sharedOrderTypes"
 
 type ExchangeError = Error & {
   statusCode: number
