+++ src/schema/v2/order/submitOrderMutation.ts
@@ -2,7 +2,7 @@ import { GraphQLString, GraphQLNonNull, GraphQLID } from "graphql"
 import {
   ORDER_MUTATION_FLAGS,
   OrderMutationResponseType,
-} from "./sharedTypes/sharedOrderTypes"
+} from "./types/sharedOrderTypes"
 import { mutationWithClientMutationId } from "graphql-relay"
 import { ResolverContext } from "types/graphql"
