+++ src/schema/v2/order/unsetOrderFulfillmentOptionMutation.ts
@@ -4,7 +4,7 @@ import { ResolverContext } from "types/graphql"
 import {
   OrderMutationResponseType,
   ORDER_MUTATION_FLAGS,
-} from "./sharedTypes/sharedOrderTypes"
+} from "./types/sharedOrderTypes"
 import { handleExchangeError } from "./exchangeErrorHandling"
 
 interface Input {
