+++ src/schema/v2/order/updateOrderMutation.ts
@@ -4,7 +4,7 @@ import {
   OrderMutationResponseType,
   OrderPaymentMethodEnum,
   OrderCreditCardWalletTypeEnum,
-} from "./sharedTypes/sharedOrderTypes"
+} from "./types/sharedOrderTypes"
 import { mutationWithClientMutationId } from "graphql-relay"
 import { ResolverContext } from "types/graphql"
 import { handleExchangeError } from "./exchangeErrorHandling"
